+++ kernel/src/string.c
@@ -1,7 +1,7 @@
 #include "../lib/string.h"
 
-uint32_t strlen(char* str) {
-    uint32_t length = 0;
+uint16_t strlen(char* str) {
+    uint16_t length = 0;
     while(*str != 0) {
         str++;
         length++;
